requirements.txt
Flask==2.0.1
python-nmap==0.6.1